test(client): unify wrangler@4 handling in e2e tests via new e2e/_utils/wrangler.ts utility
